feat: Refactor order data and update related UI components
This commit refactors the order data handling and updates the related UI components to reflect the changes.

-   **Data Model Changes:**
    -   Removed `OrderHistoryResponse` and `TrackOrderResponse`.
    -   Updated `OrderResponse` to include more comprehensive order details, such as estimate price, total price, and formatted date/time strings.
    -   Updated data types for `totalPrice` in `ShopOrderResponse` and `CreateOrderResponse` to `Int`.
    -   Updated data type for `price` in `ShopDetailResponse.ShopDetailServiceResponse` and `ShopServiceResponse` to `Int`.

-   **UI Component Updates:**
    -   Updated `OrderCard` and `TrackOrderDialog` to use the new `OrderResponse` structure.
    -   Improved layout and information display in `OrderCard`, including estimate price, total price, and formatted date.
    -   Enhanced `TrackOrderDialog` to show order details more clearly, including estimate price, total price, and special instructions.
    -   Updated `TrackingStepsView` to handle more order statuses and display appropriate messages for error states (e.g., canceled, payment failed).
    -   Added preview for `StepIndicator` and `TrackOrderDialog`.

-   **ViewModel and Repository Changes:**
    -   Updated `CustomerViewModel` to use the new `OrderResponse` in `OrderHistory` and `TrackOrder` states.
    -   Modified `CustomerRepository` and its implementations to work with the updated `OrderResponse`.
    -   Adjusted fake data in `FakeCustomerRepositoryImpl` to align with the new `OrderResponse` structure.

-   **API Service and Request Changes:**
    -   Updated API endpoints in `ApiService` to reflect the changes in order data structures.
    -   Removed unnecessary headers in `ApiService`.
    -   Added authorization headers to `getOrderHistory` and `trackOrder` API calls.

-   **Other Changes:**
    -   Updated currency formatting in `CurrencyUtil` to handle `Int` type for amount.
    -   Made `ShopOwnerActivity` exported in `AndroidManifest.xml`.
    -   Improved error handling in `CustomerHomeContent` to display error messages for shop search failures.
    -   Added `PreCheckAndQuoteBottomSheet` composable.

+++ app/src/main/java/com/example/manage/laundry/data/model/response/CreateOrderResponse.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CreateOrderResponse(
     val orderId: Int,
-    val totalPrice: Double,
+    val totalPrice: Int,
     val status: String,
     val createdAt: String
 )
