+++ app/src/main/java/com/example/manage/laundry/utils/CurrencyUtil.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import com.example.manage.laundry.R
 
 @SuppressLint("DefaultLocale")
-fun formatCurrency(context: Context, amount: Double): String {
-    val formattedNumber = String.format("%,d", amount.toInt()).replace(',', '.')
+fun formatCurrency(context: Context, amount: Int): String {
+    val formattedNumber = String.format("%,d", amount).replace(',', '.')
     return context.getString(R.string.vnd_format, formattedNumber)
 }
